requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
opencv-python-headless
scikit-learn==1.7.2
Pillow
ultralytics
aqlite